+++ Block/Adminhtml/Button/Back.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Snowdog\Menu\Block\Adminhtml\Button;
+
+use Magento\Framework\View\Element\UiComponent\Control\ButtonProviderInterface;
+
+class Back extends AbstractButton implements ButtonProviderInterface
+{
+    public function getButtonData(): array
+    {
+        return [
+            'label' => __('Back'),
+            'on_click' => sprintf("location.href = '%s';", $this->getUrl('*/*/')),
+            'class' => 'back',
+            'sort_order' => 10
+        ];
+    }
+}